requirements.txt
asgiref==3.8.1
click==8.1.7
Django==5.1.2
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
environ==1.0
inflection==0.5.1
packaging==24.1
pillow==11.0.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
setuptools==75.3.0
sqlparse==0.5.1
svgwrite==1.4.3
Tree==0.2.4
uritemplate==4.1.1
whitenoise==6.8.2
